pyproject.toml
[project]
name = "sentinelnet"
version = "0.1.0"
description = "PyTorch-based Network Intrusion Detection with Adversarial Robustness Research"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Rafael Garcia", email = "rafael.garcia.contact.me@gmail.com"}
]
requires-python = ">=3.10"
keywords = ["pytorch", "intrusion-detection", "adversarial-ml", "security", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "datasets>=2.14.0",  # HuggingFace datasets
    "huggingface-hub>=0.17.0",
    "wandb>=0.15.0",
    "optuna>=3.3.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]
adversarial = [
    "cleverhans>=4.0.0",
    "adversarial-robustness-toolbox>=1.15.0",
]

[project.scripts]
sentinelnet-train = "training.train:main"
sentinelnet-serve = "inference.serve:main"
sentinelnet-evaluate = "adversarial.evaluate_robustness:main"

[project.urls]
Homepage = "https://github.com/jag18729/sentinelnet"
Documentation = "https://github.com/jag18729/sentinelnet#readme"
Repository = "https://github.com/jag18729/sentinelnet"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["data*", "preprocessing*", "models*", "adversarial*", "training*", "inference*", "monitoring*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing"
